+++ common/components/ActivityItem/ActivityItem.tsx
@@ -1,5 +1,7 @@
-import { openActivityModalAtom, activityAtom } from '@atoms/activity-dialog';
+import { activityAtom } from '@atoms/activity-dialog';
 import { activityReloader } from '@atoms/reloaders';
+import { MOBILE_BREAKPOINT, screenWidthAtom } from '@atoms/screen';
+import ActivityDialog from '@commonComponents/ActivityDialog/ActivityDialog';
 import Button from '@commonComponents/Button/Button';
 import Dialog from '@commonComponents/Dialog/Dialog';
 import { findIconDefinition } from '@fortawesome/fontawesome-svg-core';
@@ -12,7 +14,7 @@ import { getActivityValue } from '@utils/activity-utils';
 import { getDateInputFormatFromDate } from '@utils/date';
 import { useAddNotification } from '@utils/notifications';
 import classNames from 'classnames';
-import { useSetAtom } from 'jotai';
+import { useAtomValue, useSetAtom } from 'jotai';
 import { useState } from 'react';
 import styles from './ActivityItem.module.scss';
 
@@ -20,11 +22,12 @@ interface ActivityItemProps {
   activity: Activity;
 }
 export default function ActivityItem(props: ActivityItemProps) {
-  const setOpenActivityModal = useSetAtom(openActivityModalAtom);
   const setActivity = useSetAtom(activityAtom);
   const [deleteDialogOpen, setDeleteDialogOpen] = useState(false);
   const addNotification = useAddNotification();
   const setReloader = useSetAtom(activityReloader);
+  const screenWidth = useAtomValue(screenWidthAtom);
+  const [openActivityModal, setOpenActivityModal] = useState(false);
 
   const deleteActivity = async () => {
     try {
@@ -39,6 +42,11 @@ export default function ActivityItem(props: ActivityItemProps) {
 
   return (
     <>
+      <ActivityDialog
+        openActivityModal={openActivityModal}
+        handleClose={() => setOpenActivityModal(false)}
+      ></ActivityDialog>
+
       <div className={classNames(styles.item)}>
         <div
           className={classNames(styles.header, {
@@ -55,10 +63,14 @@ export default function ActivityItem(props: ActivityItemProps) {
                 iconName: props.activity.category?.icon,
               })}
               width={14}
+              className={classNames(styles['category-icon'])}
+              title={props.activity.category?.name}
             ></FontAwesomeIcon>
           )}
           &nbsp;
-          {props.activity.category?.name}
+          <span className={classNames(styles['category-name'])}>
+            {screenWidth >= MOBILE_BREAKPOINT && props.activity.category?.name}
+          </span>
           <div className={styles.spacer}></div>
           <div className={styles.actions}>
             <Button
@@ -71,6 +83,7 @@ export default function ActivityItem(props: ActivityItemProps) {
             >
               <FontAwesomeIcon icon={faEdit} width={14}></FontAwesomeIcon>
             </Button>
+
             <Button
               fillType="regular"
               color="dark"
@@ -83,6 +96,9 @@ export default function ActivityItem(props: ActivityItemProps) {
           </div>
         </div>
         <div className={classNames(styles.body)}>
+          <span className={classNames(styles['category-name'])}>
+            {screenWidth < MOBILE_BREAKPOINT && props.activity.category?.name}
+          </span>
           <span className={classNames(styles.value)}>
             {getActivityValue(
               props.activity,