+++ pages/app/options/index.tsx
@@ -3,7 +3,6 @@ import { getAppLayout } from '@components/AppLayout/AppLayout';
 import { themeAtom } from '@atoms/theme';
 import ThemeSelector from '@components/ThemeSelector/ThemeSelector';
 import Head from 'next/head';
-import Button from '@commonComponents/Button/Button';
 
 export default function Options() {
   const [theme, setTheme] = useAtom(themeAtom);